sliit code domain core services completed
