+++ src/core/user/domain/domainServices/user.domainServices.ts
@@ -1,16 +1,14 @@
 import { Injectable } from '@nestjs/common';
 import { UserRepositories } from '../../repositories/user.repositories';
-import { registerEmailTemplate } from 'src/email-templates/register';
 import { MailOptions } from 'nodemailer/lib/smtp-transport';
 import { UserSecurity } from '../../security/user.security';
-import { USER_S3_BUCKET } from '../../user.config';
-import crypto from 'bcrypt';
-import { FORGOT_PASS_EXT, LOGIN_EXT } from '../../user.config';
 import { S3Service } from 'src/Infrastructure/s3/s3.service';
 import { User } from '@prisma/client';
-import fs from 'fs';
-import path from 'path';
-import { approveRegisterRequest, registerRequest, register } from './register.domainServices';
+import { approveRegisterRequest, registerRequest } from './requestRegister.domainServices';
+import { forgotPassword, submitForgotPassword } from './forgotPass.domainServices'
+import { login, logout, register } from './auth.domainServices'
+import { verifyTOTP, enableTOTP, disableTOTP } from './2fa.domainServices'
+import { getUser, editUser, uploadAvatar } from './profile.domainServices'
 
 
 @Injectable()
@@ -23,6 +21,9 @@ export class UserDomainServices {
     // eslint-disable-next-line no-unused-vars
     private readonly s3Service: S3Service,
   ) { }
+
+
+  // Perform register request login logout  
 
   async registerRequest(email: string): Promise<{
     isError: boolean;
@@ -53,7 +54,7 @@ export class UserDomainServices {
     message: string;
     mailOptions?: MailOptions;
   }> {
-    return  await register(data, this.userRepositories)
+    return await register(data, this.userRepositories)
   }
 
   async login(credentials: { email: string; password: string }): Promise<{
@@ -65,94 +66,23 @@ export class UserDomainServices {
     is2FAEnabled?: boolean;
     twoFAnonce?: string;
   }> {
-    try {
-      // is email valid
-      const user = await this.userRepositories.getUser(credentials.email);
-      if (!user) {
-        return {
-          isError: true,
-          message: 'Invalid password or email',
-        };
-      }
-
-      //compare password
-      const isPasswordMatch = await crypto.compare(
-        credentials.password,
-        user.password,
-      );
-
-      if (!isPasswordMatch) {
-        return {
-          isError: true,
-          message: 'Invalid password or email',
-        };
-      }
-
-      //after all valid gen access token
-      const token = this.userSecurityServices.signJWT(
-        { email: credentials.email, userId: user.id },
-        '1h',
-        'AUTHENTICATION',
-        true,
-      );
-      if (!token) {
-        throw new Error('Error signing token');
-      }
-
-      //after gen acces token gen refresh token
-      const refreshToken = this.userSecurityServices.signJWT(
-        { email: credentials.email, userId: user.id },
-        '7d',
-        'REFRESH',
-        true
-      );
-      if (!refreshToken) {
-        throw new Error('Error signing refresh token');
-      }
-
-      //if 2fa is enable
-      if (user.totpSecret && user.totpSecret !== '') {
-        // generate nonce
-        const nonce = this.userSecurityServices.genNonce();
-
-        //save nonce to redis
-        const reditsWriteResult = await this.userRepositories.writeToRedis(
-          `${user.id}${LOGIN_EXT}`,
-          JSON.stringify({
-            token,
-            nonce,
-            refreshToken,
-          }),
-          '15m',
-        );
-
-        if (!reditsWriteResult) {
-          throw new Error('Error writing nonce');
-        }
-
-        return {
-          isError: false,
-          message: 'User need to verify TOTP',
-          is2FAEnabled: true,
-          twoFAnonce: nonce,
-        };
-      }
-
-      return {
-        isError: false,
-        message: 'User logged in successfully',
-        token,
-        refreshToken,
-      };
-    } catch (error) {
-      console.log('Internal Error', error);
-      return {
-        isError: true,
-        isInternalError: true,
-        message: 'Error login',
-      };
-    }
+    return await login(credentials, this.userRepositories, this.userSecurityServices)
   }
+
+
+
+  async logout(
+    accessToken: string,
+  ): Promise<{
+    isError: boolean;
+    isInternalError?: boolean;
+    message: string;
+  }> {
+    return await logout(accessToken, this.userRepositories, this.userSecurityServices)
+  }
+
+
+  // Perform forgot password
 
   async forgotPassword(body: { email: string }): Promise<{
     isError: boolean;
@@ -160,71 +90,9 @@ export class UserDomainServices {
     message: string;
     mailOptions?: MailOptions;
   }> {
-    try {
-      const user = await this.userRepositories.getUser(body.email);
-      if (!user) {
-        return {
-          isError: true,
-          message: 'User not found',
-        };
-      }
 
-      const forgotPasswordToken = this.userSecurityServices.signJWT(
-        { email: body.email, userId: user.id },
-        '15m',
-        'FORGOT_PASSWORD',
-        true
-      );
-      if (!forgotPasswordToken) {
-        throw new Error('Error signing token');
-      }
-      const reditsWriteResult = await this.userRepositories.writeToRedis(
-        user.id + FORGOT_PASS_EXT,
-        JSON.stringify({}),
-        '15m',
-      );
+    return await forgotPassword(body, this.userRepositories, this.userSecurityServices)
 
-      if (!reditsWriteResult) {
-        return {
-          isError: true,
-          message: 'Error writing reset password token',
-        };
-      }
-
-      const emailContent = registerEmailTemplate(
-        'Your reset password request!',
-        ` 
-                You have requested to reset your password. Please click the link below to create a new password. This link will expire in 15 minutes.
-                If you did not request this, please ignore this email.
-                <br/>
-                <a href="${process.env.FRONTEND_URL}/reset-password?p=${forgotPasswordToken}"
-                style="color: rgb(0, 141, 163); --darkreader-inline-color: #5ae9ff; margin-top: 10px;"
-                target="_blank"
-                data-saferedirecturl="">Reset Link!
-               </a> 
-            `,
-      );
-
-      const mailOptions = {
-        from: 'admin@wliafdew.dev',
-        to: body.email,
-        subject: 'Register request',
-        html: emailContent,
-      };
-
-      return {
-        isError: false,
-        message: 'Forgot password email sent',
-        mailOptions,
-      };
-    } catch (error) {
-      console.log('Internal Error', error);
-      return {
-        isError: true,
-        isInternalError: true,
-        message: 'Error forgot password',
-      };
-    }
   }
 
   async submitForgotPassword(body: {
@@ -236,85 +104,12 @@ export class UserDomainServices {
     message: string;
     mailOptions?: MailOptions;
   }> {
-    try {
-      const tokenResult = await this.userSecurityServices.verifyJWT(
-        body.token,
-        'FORGOT_PASSWORD',
-      );
-      if (!tokenResult.isValid || !tokenResult.userId || !tokenResult.email) {
-        return {
-          isError: true,
-          message: 'Invalid token',
-        };
-      }
-      const isTokenUsed = await this.userRepositories.checkIsKeyIsExist(
-        tokenResult.userId + FORGOT_PASS_EXT,
-      );
-      if (isTokenUsed === null) {
-        throw new Error('Error checking token');
-      }
-
-      if (!isTokenUsed) {
-        return {
-          isError: true,
-          message: 'Token already used',
-        };
-      }
-
-      const newPass = await crypto.hash(body.password, 10);
-
-      const user = await this.userRepositories.updateUser(
-        tokenResult.email,
-        'password',
-        newPass,
-      );
-
-      if (!user) {
-        throw new Error('Error updating user password');
-      }
-
-      const removeResult = await this.userRepositories.removeKey(
-        tokenResult.userId + FORGOT_PASS_EXT,
-      );
-      const invalidTokenResult = await this.userRepositories.removeKey(
-        `${user.email}-${'FORGOT_PASSWORD'}`,
-      );
-
-      if (!invalidTokenResult) {
-        throw new Error('Error removing invalid token');
-      }
-
-      if (!removeResult) {
-        throw new Error('Error removing reset password token');
-      }
-
-      const emailContent = registerEmailTemplate(
-        'Your password has been reset!',
-        'You have successfully reset your password. You can now login to your account.',
-      );
-
-      const mailOptions = {
-        from: 'admin@wliafdew.dev',
-        to: tokenResult.email,
-        subject: 'Register request',
-        html: emailContent,
-      };
-
-      return {
-        isError: false,
-        message: 'Password reset successfully',
-        mailOptions,
-      };
-    } catch (error) {
-      console.log('Internal Error', error);
-      return {
-        isError: true,
-        isInternalError: true,
-        message: 'Error submit forgot password',
-      };
-    }
+    return await submitForgotPassword(body, this.userRepositories, this.userSecurityServices)
   }
 
+
+
+  // Perform 2FA totp
   async enableTOTP(
     email: string,
     password: string,
@@ -324,50 +119,7 @@ export class UserDomainServices {
     message: string;
     qrCodeImageURL?: string;
   }> {
-    try {
-      const user = await this.userRepositories.getUser(email);
-
-      if (!user) {
-        return {
-          isError: true,
-          message: 'User not found',
-        };
-      }
-
-      if (user?.totpSecret && user.totpSecret !== '') {
-        return {
-          isError: true,
-          message: 'User already have 2FA TOTP enable',
-        };
-      }
-
-      const isPasswordMatch = await crypto.compare(password, user.password);
-      if (!isPasswordMatch) {
-        return {
-          isError: true,
-          message: 'Invalid password',
-        };
-      }
-
-      const genTOTP = await this.userSecurityServices.generateOTP(email);
-
-      if (!genTOTP.qrCodeImageURL) {
-        throw new Error('Error generating qr code');
-      }
-
-      return {
-        isError: false,
-        message: 'TOTP enabled successfully',
-        qrCodeImageURL: genTOTP.qrCodeImageURL,
-      };
-    } catch (error) {
-      console.log('Internal Error', error);
-      return {
-        isError: true,
-        isInternalError: true,
-        message: 'Error enabling TOTP',
-      };
-    }
+    return await enableTOTP(email, password, this.userRepositories, this.userSecurityServices)
   }
 
   async disableTOTP(
@@ -378,45 +130,7 @@ export class UserDomainServices {
     isInternalError?: boolean;
     message: string;
   }> {
-    try {
-      const user = await this.userRepositories.getUser(email);
-
-      if (!user) {
-        return {
-          isError: true,
-          message: 'User not found',
-        };
-      }
-
-      if (!user?.totpSecret || user.totpSecret === '') {
-        return {
-          isError: true,
-          message: 'User does not have 2FA TOTP enable',
-        };
-      }
-
-      const isPasswordMatch = await crypto.compare(password, user.password);
-      if (!isPasswordMatch) {
-        return {
-          isError: true,
-          message: 'Invalid password',
-        };
-      }
-
-      await this.userRepositories.updateUser(email, 'totpSecret', null);
-
-      return {
-        isError: false,
-        message: 'TOTP disable successfully',
-      };
-    } catch (error) {
-      console.log('Internal Error', error);
-      return {
-        isError: true,
-        isInternalError: true,
-        message: 'Error disabling TOTP',
-      };
-    }
+    return await disableTOTP(email, password, this.userRepositories)
   }
 
   async verifyTOTP(
@@ -430,130 +144,11 @@ export class UserDomainServices {
     token?: string;
     refreshToken?: string;
   }> {
-    try {
-      const user = await this.userRepositories.getUser(email);
-
-      if (!user) {
-        return {
-          isError: true,
-          message: 'User not found',
-        };
-      }
-
-      if (!user?.totpSecret || user.totpSecret === '') {
-        return {
-          isError: true,
-          message: 'User does not have 2FA TOTP enable',
-        };
-      }
-
-      const cachingLogin = await this.userRepositories.getValueFromRedis(
-        `${user?.id}${LOGIN_EXT}`,
-      );
-      if (!cachingLogin) {
-        return {
-          isError: true,
-          message: 'Nonce not found',
-        };
-      }
-
-      if (cachingLogin.nonce !== nonce) {
-        return {
-          isError: true,
-          message: 'Invalid nonce',
-        };
-      }
-
-      if (!cachingLogin.token || !cachingLogin.refreshToken) {
-        throw new Error('Error getting token');
-      }
-
-      const verifyResult = this.userSecurityServices.verifyOTP(
-        email,
-        token,
-        user,
-      );
-
-      if (verifyResult.isError) {
-        return {
-          isError: true,
-          message: verifyResult.message,
-        };
-      }
-
-      if (verifyResult.isInterNalError) {
-        throw new Error('Error verifying OTP');
-      }
-
-      return {
-        isError: false,
-        message: 'Access granted TOTP verified successfully',
-        refreshToken: cachingLogin.refreshToken,
-        token: cachingLogin.token,
-      };
-    } catch (error) {
-      console.log('Internal Error', error);
-      return {
-        isError: true,
-        isInternalError: true,
-        message: 'Error verifying TOTP',
-      };
-    }
+    return await verifyTOTP(email, token, nonce, this.userRepositories, this.userSecurityServices)
   }
 
 
 
-  async logout(
-    accessToken: string,
-  ): Promise<{
-    isError: boolean;
-    isInternalError?: boolean;
-    message: string;
-  }> {
-    try {
-      //verify access token
-
-      const verifyAccessTokenResult = await this.userSecurityServices.verifyJWT(
-        accessToken,
-        'AUTHENTICATION'
-      )
-
-      if (!verifyAccessTokenResult.isValid) {
-        return {
-          isError: true,
-          message: 'Invalid access token',
-        }
-      }
-
-      //invalid refresh token from redis
-      const invalidResult = await this.userRepositories.removeKey(
-        `${verifyAccessTokenResult.email}-REFRESH`
-      )
-
-      //invalid access token from redis
-      const invalidAccessToken = await this.userRepositories.removeKey(
-        `${verifyAccessTokenResult.email}-AUTHENTICATION`
-      )
-
-      if (!invalidResult || !invalidAccessToken) {
-        throw new Error('Error removing refresh token')
-      }
-
-      // all process success
-      return {
-        isError: false,
-        message: 'Logout successfully',
-      };
-
-    } catch (error) {
-      console.log('Internal Error', error);
-      return {
-        isError: true,
-        isInternalError: true,
-        message: 'Error during logout',
-      };
-    }
-  }
 
   async getUser(
     userId: string,
@@ -563,41 +158,7 @@ export class UserDomainServices {
     message: string;
     user?: User;
   }> {
-    try {
-
-      // get user from database 
-      const userData = await this.userRepositories.getUser('_', userId);
-
-      if (!userData) {
-        return {
-          isError: true,
-          message: 'User not found',
-        };
-      }
-
-
-      //remove sensitive data
-      const userDataClone = { ...userData };
-      delete (userDataClone as Partial<User>).password;
-      delete (userDataClone as Partial<User>).totpSecret;
-
-
-
-
-      return {
-        isError: false,
-        message: 'User retrieved successfully',
-        user: userDataClone,
-      };
-
-    } catch (error) {
-      console.log('Internal Error', error);
-      return {
-        isError: true,
-        isInternalError: true,
-        message: 'Error retrieving user',
-      };
-    }
+    return await getUser(userId, this.userRepositories)
   }
 
 
@@ -610,38 +171,8 @@ export class UserDomainServices {
     isInternalError?: boolean;
     message: string;
   }> {
-    try {
 
-      for (const key in data) {
-        if (!data[key]) {
-          delete data[key];
-        }
-      }
-
-
-      // get user from database 
-      const userData = await this.userRepositories.updateUser('_', '_', '_', userId, true, data as User);
-
-      if (!userData) {
-        return {
-          isError: true,
-          message: 'User not found',
-        };
-      }
-
-      return {
-        isError: false,
-        message: 'User retrieved successfully',
-      };
-
-    } catch (error) {
-      console.log('Internal Error', error);
-      return {
-        isError: true,
-        isInternalError: true,
-        message: 'Error retrieving user',
-      };
-    }
+    return await editUser(userId, data, this.userRepositories)
   }
 
 
@@ -652,70 +183,7 @@ export class UserDomainServices {
     isInternalError?: boolean;
     message: string;
   }> {
-    try {
-      //verify is user exist
 
-      const user = await this.userRepositories.getUser('_', userId);
-
-
-      if (!user) {
-        return {
-          isError: true,
-          message: 'Invalid access token',
-        }
-      }
-
-
-      //perform upload to s3
-
-      const newAvatar = `${user.id}/avatar/avatar${path.extname(name)}`
-      const readStream = fs.createReadStream(avatar)
-      await this.s3Service.upLoadToS3(USER_S3_BUCKET, newAvatar, readStream)
-
-      //clean up old avatar on s3 this will be not efficient if the avatar is large so i run this in background
-      //if the error happen i need cached to redis and crond job to clean up the old avatar
-      this.s3Service.s3.listObjectsV2({
-        Bucket: USER_S3_BUCKET,
-        Prefix: `${user.id}/avatar/`,
-        MaxKeys: 1000
-      }, (err, data) => {
-        if (err) {
-          console.log("Internal Error", err)
-          return
-        }
-        if (!data.Contents) return
-        for (const content of data.Contents) {
-          if (content.Key === newAvatar) continue
-          this.s3Service.s3.deleteObject({
-            Bucket: USER_S3_BUCKET,
-            Key: content.Key!
-          }, (err) => {
-            if (err) {
-              console.log("Internal Error", err)
-              return
-            }
-          })
-          console.log(`Deleted ${content.Key}`)
-        }
-      })
-
-
-
-
-      return {
-        isError: false,
-        message: 'Avatar uploaded successfully',
-      };
-
-
-
-    } catch (error) {
-      console.log('Internal Error', error);
-      return {
-        isError: true,
-        isInternalError: true,
-        message: 'Error uploading user avatar',
-      };
-    }
+    return await uploadAvatar(userId, avatar, name, this.userRepositories, this.s3Service)
   }
 }
